Update the user test.
