+++ tests/AppBundle/Utils/User/ForgotPasswordTypeHandlerTest.php
@@ -3,7 +3,9 @@
 namespace tests\AppBundle\Utils;
 
 use AppBundle\Entity\User;
+use Doctrine\ORM\QueryBuilder;
 use Symfony\Component\Form\Form;
+use Doctrine\ORM\EntityRepository;
 use AppBundle\Repository\UserRepository;
 use AppBundle\Form\User\ForgotPasswordType;
 use Symfony\Component\Form\Test\TypeTestCase;
@@ -33,23 +35,45 @@ class ForgotPasswordTypeHandlerTest extends TypeTestCase
      * @var Session
      */
     private $session;
+    
+    /**
+     * @var \Twig_Environment
+     */
+    private $twig;
 
     /**
-     * @var EventDispatcherInterface
+     * @var \Twig_Environment
      */
-    private $eventDispatcher;
+    private $template;
+
+    /**
+     * @var \Swift_Mailer
+     */
+    private $mailer;
 
     protected function setUp()
     {
         parent::setUp();
 
         // Last, mock the EntityManager to return the mock of the repository
-        $this->entityManager = $this->createMock(ObjectManager::class);
+        $this->entityManager = $this
+            ->getMockBuilder(ObjectManager::class)
+            ->disableOriginalConstructor()
+            ->getMock();
 
         $this->session = new Session(new MockArraySessionStorage());
 
-        $this->eventDispatcher = $this
-            ->getMockBuilder(EventDispatcherInterface::class)
+        $this->mailer = $this
+        ->getMockBuilder(\Swift_Mailer::class)
+        ->disableOriginalConstructor()
+        ->getMock();
+        $this->template = $this
+            ->getMockBuilder(\Twig_Template::class)
+            ->disableOriginalConstructor()
+            ->setMethods(['renderBlock'])
+            ->getMockForAbstractClass();
+        $this->twig = $this
+            ->getMockBuilder(\Twig_Environment::class)
             ->disableOriginalConstructor()
             ->getMock();
     }
@@ -60,16 +84,29 @@ class ForgotPasswordTypeHandlerTest extends TypeTestCase
         $userRepository = $this
             ->getMockBuilder(UserRepository::class)
             ->disableOriginalConstructor()
-            ->setMethods(['findOneByEmail'])
+            ->setMethods(['getUserWithEmail'])
             ->getMock();
         $userRepository
             ->expects($this->once())
-            ->method('findOneByEmail')
+            ->method('getUserWithEmail')
             ->willReturn(new User());
 
         $this->entityManager->expects($this->any())
             ->method('getRepository')
             ->willReturn($userRepository);
+
+        $this->mailer
+            ->expects($this->once())
+            ->method('send')
+            ->willReturn(true);
+        $this->template
+            ->expects($this->any())
+            ->method('renderBlock')
+            ->willReturn('test');
+        $this->twig
+            ->expects($this->once())
+            ->method('load')
+            ->willReturn($this->template);
 
         $formData = [
             'emailRecovery' => 'email@test.com',
@@ -83,23 +120,73 @@ class ForgotPasswordTypeHandlerTest extends TypeTestCase
         $handler = new ForgotPasswordTypeHandler(
             $this->entityManager,
             $this->session,
-            $this->eventDispatcher
+            $this->twig,
+            $this->mailer
         );
 
         $this->assertTrue($handler->handle($form));
     }
 
-    public function testHandleFashBecauseNoUser()
+    public function testErroSsendingEmail()
     {
         // Now, mock the repository so it returns the mock of the user
         $userRepository = $this
             ->getMockBuilder(UserRepository::class)
             ->disableOriginalConstructor()
-            ->setMethods(['findOneByEmail'])
+            ->setMethods(['getUserWithEmail'])
             ->getMock();
         $userRepository
             ->expects($this->once())
-            ->method('findOneByEmail')
+            ->method('getUserWithEmail')
+            ->willReturn(new User());
+
+        $this->entityManager->expects($this->any())
+            ->method('getRepository')
+            ->willReturn($userRepository);
+
+        $this->mailer
+            ->expects($this->once())
+            ->method('send')
+            ->willReturn(false);
+        $this->template
+            ->expects($this->any())
+            ->method('renderBlock')
+            ->willReturn('test');
+        $this->twig
+            ->expects($this->once())
+            ->method('load')
+            ->willReturn($this->template);
+
+        $formData = [
+            'emailRecovery' => 'email@test.com',
+        ];
+
+        $form = $this->factory->create(ForgotPasswordType::class);
+
+        // submit the data to the form directly
+        $form->submit($formData);
+
+        $handler = new ForgotPasswordTypeHandler(
+            $this->entityManager,
+            $this->session,
+            $this->twig,
+            $this->mailer
+        );
+
+        $this->assertTrue($handler->handle($form));
+    }
+
+    public function testHandleFalseBecauseNoUser()
+    {
+        // Now, mock the repository so it returns the mock of the user
+        $userRepository = $this
+            ->getMockBuilder(UserRepository::class)
+            ->disableOriginalConstructor()
+            ->setMethods(['getUserWithEmail'])
+            ->getMock();
+        $userRepository
+            ->expects($this->once())
+            ->method('getUserWithEmail')
             ->willReturn(null);
 
         $this->entityManager->expects($this->any())
@@ -118,7 +205,8 @@ class ForgotPasswordTypeHandlerTest extends TypeTestCase
         $handler = new ForgotPasswordTypeHandler(
             $this->entityManager,
             $this->session,
-            $this->eventDispatcher
+            $this->twig,
+            $this->mailer
         );
 
         $this->assertFalse($handler->handle($form));
@@ -138,7 +226,8 @@ class ForgotPasswordTypeHandlerTest extends TypeTestCase
         $handler = new ForgotPasswordTypeHandler(
             $this->entityManager,
             $this->session,
-            $this->eventDispatcher
+            $this->twig,
+            $this->mailer
         );
 
         $this->assertFalse($handler->handle($form));
